mysql: Convert known errors to GRPC status (#2331)
This helps clients have a better signal on whether the operation is restartable.
The Backoff package automatically honours the Aborted code as eligible
for restarts.

This change mitigates a flaky TestInProcessLogIntegration.
